requirements.txt
aiofile==3.8.8
aioshutil==1.3
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
Brotli==1.1.0
caio==0.9.13
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflate64==1.0.0
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
multivolumefile==0.2.3
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psutil==5.9.7
py7zr==0.20.8
pyasn1==0.5.1
pybcj==1.0.2
pycryptodomex==3.19.0
pydantic==1.10.2
pydantic-core==2.14.5
pyppmd==1.1.0
pytest==7.4.3
pytest-asyncio==0.23.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyzstd==0.15.9
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
texttable==1.7.0
tomli==2.0.1
typing-extensions==4.9.0
uvicorn==0.24.0.post1
